backend/requirements.txt
Flask==2.2.2
numpy==1.23.3
umap_learn==0.5.3
flask-cors==3.0.10
